% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_csv_tide.R
\name{cre_csv_tide}
\alias{cre_csv_tide}
\title{Orders HKOtide results}
\usage{
cre_csv_tide(DIR, load = F)
}
\arguments{
\item{DIR}{Path of the directory. Accepts list of directory.}

\item{load}{Writes CSV file to Working Directory.}
}
\value{

}
\description{
Tidies up HKOtide csv files.
Can create csv in working directory with default name.
}
\examples{
HKOtide(getwd())
}
