% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_svg_tide.R
\name{cre_svg_tide}
\alias{cre_svg_tide}
\title{Creates graph of HKOtide results (from tidy csv)}
\usage{
cre_svg_tide(df, DIR)
}
\arguments{
\item{df}{Dataframe of input data. Must contain the following columns, "Station", "Date", "Min_00...Min_55"}

\item{DIR}{Path of the directory. Accepts list of directory.}
}
\value{

}
\description{
Creates a list with ggplot graph and station name.
}
\examples{
cre_svg_tide(df = df)
}
