% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_hodo_plot.R
\name{draw_hodo_plot}
\alias{draw_hodo_plot}
\title{Draw hodograph}
\usage{
draw_hodo_plot(
  data,
  limit = c(0, 99999),
  type = "smooth",
  space = 500,
  df = 0.7,
  hght = seq(0, 99999, 1)
)
}
\arguments{
\item{data}{Data from 1 station. Only accepts a dataframe that includes column PRES, TEMP, and DWPT.}

\item{limit}{The layer of atmosphere it draws. The default is to practically draw all available data. Only accepts a list of 2 numbers.}

\item{type}{The type of hodograph drawn. Only accepts "smooth"/ "raw".}

\item{space}{Drawn when type is equal to "smooth". The spacing between each dot. The default is to draw a dot per 500 unit of altitude. Only accepts 1 number.}

\item{df}{Drawn when type is equal to "smooth". Degree of freedom. Default is set at 0.7, i.e. 70\% of all sounding data available.}

\item{hght}{Drawn when type is equal to "smooth". Height of the atmosphere to be predicted. Only accepts a list of numbers.}
}
\value{

}
\description{
Draw hodograph
}
\examples{
draw_hodo_plot(data)
}
