% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_hodo_smooth.R
\name{draw_hodo_smooth}
\alias{draw_hodo_smooth}
\title{Draws a smoothed hodograph line}
\usage{
draw_hodo_smooth(data, df = 0.7, hght = seq(0, 16000, 1))
}
\arguments{
\item{data}{Data from 1 station. Only accepts a dataframe that includes column HGHT, UWND, and VWND.}

\item{df}{Degree of freedom. Default is set at 0.7, i.e. 70\% of all sounding data available.}

\item{hght}{Height of the atmosphere to be predicted. Only accepts a list of numbers.}
}
\value{

}
\description{
Draws a smoothed hodograph line
}
\examples{
draw_hodo_smooth(data)
}
