% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_line_dabt.R
\name{draw_line_dabt}
\alias{draw_line_dabt}
\title{Loads dry adiabetic line}
\usage{
draw_line_dabt(temp = seq(-50, 50, 10), pres = c(5, 1500))
}
\arguments{
\item{temp}{Temperature of said gas at a pressure of 1000hPa.}

\item{pres}{Pressure range to draw. Default to be between 1500 to 5 hPa. Must be a list of 2 number that can be divided by 5.}
}
\value{

}
\description{
Dry diabetic line refers to the temperature of dry, moisture free gas raising into the atmosphere.
This function creates these lines by creating a data frame.
}
\examples{
draw_line_dabt()
}
