% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_line_ihum.R
\name{draw_line_ihum}
\alias{draw_line_ihum}
\title{Draw isohume lines}
\usage{
draw_line_ihum(rs = c(0.1, 0.2, 0.5, 1, 2, 5, 10, 30, 50), pres = c(1500, 5))
}
\arguments{
\item{rs}{Mixing ratios to be drawn, in g/kg.}

\item{pres}{Pressure range to draw. Default to be between 1500 to 5 hPa. Must be a list of 2 number that can be divided by 5.}
}
\value{

}
\description{
Isohume lines refers to lines of the same mixing ratio, at different temperature and pressure (assuming the air is saturated).
This function creates these lines by creating a data frame.
}
\examples{
draw_line_ihum()
}
