% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_line_mabt.R
\name{draw_line_mabt}
\alias{draw_line_mabt}
\title{Loads moist adiabetic line}
\usage{
draw_line_mabt(temp = seq(-50, 50, 10))
}
\arguments{
\item{temp}{Temperature of said gas at a pressure of 1000hPa.}
}
\value{

}
\description{
Moist diabetic line refers to the temperature of saturated gas raising into the atmosphere.
This function creates these lines by creating a data frame.
}
\examples{
load_line_mabt()
}
