% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_dabt.R
\name{enq_dabt}
\alias{enq_dabt}
\title{Find condition, dry-adibetically}
\usage{
enq_dabt(temp1, pres1, pres2, temp2, find = "temp")
}
\arguments{
\item{temp1}{Air temperature at condition 1, in degree Celsius}

\item{pres1}{Air pressure at condition 1, in hPa}

\item{pres2}{Air pressure at condition 2, in hPa}

\item{temp2}{Air temperature at condition 2, in degree Celsius}

\item{find}{what to find? temp or pres?}
}
\value{

}
\description{
Find condition, dry-adibetically
}
\examples{
enq_dabt(20, 1000, 800, find = "temp")
}
