% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_es.R
\name{enq_es}
\alias{enq_es}
\title{Find the saturation vapor pressure or reverse}
\usage{
enq_es(temp, es, find = "es")
}
\arguments{
\item{temp}{Air temperature, in degree Celsius}

\item{es}{Saturation vapor pressure, in hPa}

\item{find}{what to find}
}
\value{

}
\description{
Using Buck equation. Check more in https://www.omnicalculator.com/chemistry/vapour-pressure-of-water.
}
