% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_ihum.R
\name{enq_ihum}
\alias{enq_ihum}
\title{Find temperature and pressure at isohume or reverse, assuming saturated}
\usage{
enq_ihum(temp1, pres1, pres2, temp2, find = "temp")
}
\arguments{
\item{temp1}{Air temperature at condition 1, in degree Celsius}

\item{pres1}{Air pressure at condition 1, in hPa}

\item{pres2}{Air pressure at condition 2, in hPa}

\item{temp2}{Air temperature at condition 2, in degree Celsius}

\item{find}{what to find? temp or pres?}
}
\value{

}
\description{
Find temperature and pressure at isohume or reverse, assuming saturated
}
\examples{
enq_ihum(0, 1013, 500, find = "temp")
}
