% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_lcl_est.R
\name{enq_lcl_est}
\alias{enq_lcl_est}
\title{Find lifted condensation level (LCL) through estimation}
\usage{
enq_lcl_est(dwpt, temp, pres)
}
\arguments{
\item{dwpt}{Dewpoint temperature, in degree Celsius}

\item{temp}{Air temperature, in degree Celsius}

\item{pres}{Air pressure, in hPa}
}
\value{

}
\description{
Only outputs lifted condensation level, and not in reverse.
Only accepts dewpoint and temperature input at the same pressure.
Outputs lclp (LCL, pressure) and lclt (LCL, temperature), in the unit of hPa and deg Celsius.
}
\examples{
enq_lcl(c(10,21), 20, 1000)
}
