% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_mabt.R
\name{enq_mabt}
\alias{enq_mabt}
\title{Find condition, moist-adibetically}
\usage{
enq_mabt(temp1, pres1, pres2, acc = 0.1)
}
\arguments{
\item{temp1}{Temperature at Pressure level 1, in degC. Only accepts numbers or a list of numbers.}

\item{pres1}{Pressure at Pressure level 1, in hPa. Only accepts numbers or a list of numbers.}

\item{pres2}{Pressure at Pressure level 2, in hPa. Only accepts numbers or a list of numbers.}

\item{acc}{Accuracy of the calculation. Smaller number, Higher accuracy. Only accepts number.}
}
\value{

}
\description{
Find condition, moist-adibetically
}
\examples{
enq_mabt(c(10, 20, 30), 1000, 500) should return -125.4218, -119.2539, -112.3171, as a list
}
