% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_rs.R
\name{enq_rs}
\alias{enq_rs}
\title{Find saturation mixing ratio or reverse}
\usage{
enq_rs(temp, pres, rs, find = "rs")
}
\arguments{
\item{temp}{Air temperature, in degree Celsius}

\item{pres}{Saturation vapor pressure, in hPa}

\item{rs}{Saturation mixing ratio, in g_H2O/kg_air}

\item{find}{what to find}
}
\value{

}
\description{
Find saturation mixing ratio or reverse
}
\examples{
enq_rs(1, 1013, find = "rs")
}
