% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_temp.R
\name{enq_temp}
\alias{enq_temp}
\title{Find temperature at different common scales}
\usage{
enq_temp(degc, degf, degk)
}
\arguments{
\item{degc}{Temperature in degree Celsius}

\item{degf}{Temperature in degree Fahrenheit}

\item{degk}{Temperature in degree Kelvin}
}
\value{

}
\description{
Find temperature at different common scales
}
\examples{
enq_temp(10, -40, 273.15)
}
