% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_wind.R
\name{enq_wind}
\alias{enq_wind}
\title{Find wind speed at different common scales and tropical cyclone scales}
\usage{
enq_wind(kmh, mps, mph, fps, kt)
}
\arguments{
\item{kmh}{Wind speed in kilometer per hour}

\item{mps}{Wind speed in meter per second}

\item{mph}{Wind speed in mile per hour}

\item{fps}{Wind speed in feet per second}

\item{kt}{Wind speed in knots}
}
\value{
1 dataframe
}
\description{
Find wind speed at different common scales and tropical cyclone scales
}
\examples{
enq_wind(10, 20, 30, 40, 60)
}
