% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKOall.R
\name{HKOall}
\alias{HKOall}
\title{Downloads all the satellite images and lighting images from the Hong Kong Observatory. If there is a Chinese and English version, only downloads the English version.}
\usage{
HKOall()
}
\value{

}
\description{
Downloads all the satellite images and lighting images from the Hong Kong Observatory. If there is a Chinese and English version, only downloads the English version.
}
\examples{
HKOall()
}
