% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKOlighting.R
\name{HKOlighting}
\alias{HKOlighting}
\title{Download lighting and radar images from HKO and stores them in the working directory orderly according to date.}
\usage{
HKOlighting(
  range = 64,
  type = "CC",
  DDays = 4.5,
  DTime = Sys.time(),
  listfail = F
)
}
\arguments{
\item{range}{The radar image range in kilometers. Only accepts values 64 or 256.}

\item{type}{The type of lighting data, which includes "cloud to cloud"(CC) and "cloud to ground"(CG). Only accepts CC and CG}

\item{DDays}{The duration of images to be downloaded in days. Only accepts numerical values.}

\item{DTime}{The starting time of images to be downloaded, starting from the latest time. Only accepts POSIXct (tip: create POSIXct via base::ISOdate).}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download lighting and radar images from HKO and stores them in the working directory orderly according to date.
}
\examples{
HKOlighting()
}
