% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKOrfdy.R
\name{HKOrfdy}
\alias{HKOrfdy}
\title{Download daily rainfall images}
\usage{
HKOrfdy(ETime = Sys.time(), lan = "e", listfail = F)
}
\arguments{
\item{ETime}{The latest/ending time of images to be downloaded, at the latest time. Only accepts POSIXct (tip: create POSIXct via ISOdate).}

\item{lan}{Language of descriptive text within the pictures. Only accepts "e"/ "c" as char.}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download from HKO.
Stores them in the working directory orderly according to year.
}
\examples{
HKOrfdy()
}
