% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKOsatellite.R
\name{HKOsatellite}
\alias{HKOsatellite}
\title{Download satellite images (imaging over East Asia) from HKO and stores them in the working directory orderly according to date.}
\usage{
HKOsatellite(
  magn = 8,
  type = "dc",
  DDays = 3.5,
  DTime = Sys.time(),
  listfail = F
)
}
\arguments{
\item{magn}{The satellite image magnification. Only accepts 2/ 4/ 8}

\item{type}{The type of satellite image, which includes "True color" (tc), "Infrared" (ir), and "Deep convection" (dc). Only accepts "tc", "ir", or "dc".}

\item{DDays}{The duration of images to be downloaded in days. Only accepts numerical values.}

\item{DTime}{The starting time of images to be downloaded, starting from the latest time. Only accepts POSIXct (tip: create POSIXct via base::ISOdate).}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download satellite images (imaging over East Asia) from HKO and stores them in the working directory orderly according to date.
}
\examples{
HKOsatellite()
}
