% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKOtide.R
\name{HKOtide}
\alias{HKOtide}
\title{Download tidal data from HKO}
\usage{
HKOtide(ETime = Sys.time(), DDays = 7, STime = NA, lan = "e", listfail = F)
}
\arguments{
\item{ETime}{Ending Time. Only accepts POSIXct date/time only.}

\item{DDays}{Duration of Time between Ending Time and Starting Time (in days). Defaults as 7 days. Only accepts numerical values.}

\item{STime}{Starting Time. Only accepts POSIXct date/time only.}

\item{lan}{Language of descriptive text within the pictures. Only accepts "e"/ "c" as char.}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Downloads and stores them in the working directory orderly according to date.
This is a Level 3 code.
}
\examples{
HKOtide()
}
