% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKOwind.R
\name{HKOwind}
\alias{HKOwind}
\title{Download wind data from HKO and stores them in the working directory orderly according to date.}
\usage{
HKOwind(type = "wind", lan = "e", listfail = F)
}
\arguments{
\item{type}{Type of wind. Only accepts "gust"/ "wind" as char.}

\item{lan}{Language of descriptive text within the pictures. Only accepts "e"/ "c" as char.}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download wind data from HKO and stores them in the working directory orderly according to date.
}
\examples{
HKOwind()
}
