% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKOwind_csv.R
\name{HKOwind_csv}
\alias{HKOwind_csv}
\title{Download Wind and Gust data from HKO in CSV format}
\usage{
HKOwind_csv(ETime = Sys.time(), DDays = 7, STime = NA, listfail = F)
}
\arguments{
\item{ETime}{Ending Time. Only accepts POSIXct date/time only.}

\item{DDays}{Duration of Time between Ending Time and Starting Time (in days). Defaults as 7 days. Only accepts numerical values.}

\item{STime}{Starting Time. Only accepts POSIXct date/time only.}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Downloads and stores them in the working directory orderly according to date.
This is a Level 3 code.
}
\examples{
HKOwind_csv()
}
