% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkw_dir.cre.R
\name{hkw_dir.cre}
\alias{hkw_dir.cre}
\title{Warning! This is an internal command!
Users should not use this!
Creates directory for HKWeather. This should not be used alone.}
\usage{
hkw_dir.cre(dir = getwd(), pri, sec)
}
\arguments{
\item{dir}{xxx}

\item{pri}{xxx}

\item{sec}{xxx}
}
\value{

}
\description{
Warning! This is an internal command!
Users should not use this!
Creates directory for HKWeather. This should not be used alone.
}
\examples{
hkw_dir.cre(pri = c("Rfhr", "C"), sec = data.frame$Date)
}
