% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkw_dir.cre2.R
\name{hkw_dir.cre2}
\alias{hkw_dir.cre2}
\title{Warning! This is an internal command!
Users should not use this!
Creates directory for HKWeather. This should not be used alone.}
\usage{
hkw_dir.cre2(DIR = getwd(), expand = "all T", ...)
}
\arguments{
\item{DIR}{xxx}

\item{expand}{xxx}

\item{...}{xxx}
}
\value{

}
\description{
Warning! This is an internal command!
Users should not use this!
Creates directory for HKWeather. This should not be used alone.
}
\examples{
hkw_dir.cre2(a = c("1"))
}
