% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkw_dir.cre3.R
\name{hkw_dir.cre3}
\alias{hkw_dir.cre3}
\title{Warning! This is an internal command!
Users should not use this!
Creates directory for HKWeather. This should not be used alone.}
\usage{
hkw_dir.cre3(wDIR, filename)
}
\arguments{
\item{wDIR}{xxxx}

\item{filename}{xxxx}
}
\value{

}
\description{
Warning! This is an internal command!
Users should not use this!
Creates directory for HKWeather. This should not be used alone.
}
\examples{
hkw_dir.cre3(wDIR = paste0(getwd(), "/a", "/b", "/c.csv"), filename = T)
}
