% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkw_fil.cre2.R
\name{hkw_fil.cre2}
\alias{hkw_fil.cre2}
\title{Warning! This is an internal command!
Users should not use this!
Creates files for HKWeather.}
\usage{
hkw_fil.cre2(URL, DIR, Time, listfail)
}
\arguments{
\item{URL}{xxx}

\item{DIR}{xxx}

\item{Time}{xxx}

\item{listfail}{xxx}

\item{DTime}{xxx}

\item{DDays}{xxx}
}
\value{

}
\description{
Warning! This is an internal command!
Users should not use this!
Creates files for HKWeather.
}
\examples{
hkw_fil.cre2(URL = "1", DIR = "1.png", Time = 1, listfail = F)
}
