% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_gif.R
\name{img_gif}
\alias{img_gif}
\title{Creates a .gif file based on images within working directory.}
\usage{
img_gif(DIR = NA, name = NA, fps = 25)
}
\arguments{
\item{DIR}{Path of the directory. Accepts list of directory.}

\item{name}{Name of the gif file created. Default as the combination of start and end frame file names.}

\item{fps}{Frame per second of the gif file created. Default as 25 fps.}
}
\value{

}
\description{
Creates a .gif file based on images within working directory.
}
\examples{
img_gif()
}
