% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all.R
\name{load_all}
\alias{load_all}
\title{Downloads all the satellite images and lighting images from the Hong Kong Observatory. If there is a Chinese and English version, only downloads the English version.}
\usage{
load_all(listfail = F)
}
\value{

}
\description{
Downloads all the satellite images and lighting images from the Hong Kong Observatory. If there is a Chinese and English version, only downloads the English version.
}
\examples{
load_all()
}
