% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_asnd.R
\name{load_asnd}
\alias{load_asnd}
\title{xx}
\usage{
load_asnd(ETime = Sys.time(), DDays = 7, STime = NA)
}
\arguments{
\item{ETime}{The time of the lastest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}
}
\value{

}
\description{
xx
}
\examples{
load_asnd()
}
