% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_asnd_data.R
\name{load_asnd_data}
\alias{load_asnd_data}
\title{Download atmospheric sounding data metrics}
\usage{
load_asnd_data(ETime = Sys.time(), DDays = 7, STime = NA, Station = 45004)
}
\arguments{
\item{ETime}{The time of the lastest data. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{DDays}{The duration of data to be downloaded in days. Only accepts numerical values.}

\item{STime}{The time of the oldest data. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}
}
\value{

}
\description{
Understanding the sounding is important for understanding the atmospheric condition, especially for extreme weather.
This catches the metrics provided by the University of Wyoming.
}
\examples{
load_asnd_data()
}
