% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ltng.R
\name{load_ltng}
\alias{load_ltng}
\title{Download lighting and radar images}
\usage{
load_ltng(
  ETime = Sys.time(),
  DDays = 4.5,
  STime = NA,
  type = "cc",
  range = 64,
  listfail = F
)
}
\arguments{
\item{ETime}{The time of the lastest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{DDays}{The duration of images to be downloaded in days. Only accepts numerical values.}

\item{STime}{The time of the earliest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{type}{The type of lighting data, which includes "cloud to cloud"(CC) and "cloud to ground"(CG)}

\item{range}{The radar image range in kilometers. Only accepts values 64 or 256.}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download from HKO and stores them in the working directory orderly according to date.
}
\examples{
load_ltng()
}
