% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oisst.R
\name{load_oisst}
\alias{load_oisst}
\title{Download NOAA OISST data}
\usage{
load_oisst(EDay = "1981-09-01", SDay = "1981-09-01", listfail = F)
}
\arguments{
\item{EDay}{Ending Day, which is the later time}

\item{SDay}{Starting Day, which is the older time}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
OISST is a daily, global sea surface temperature data, with a resolution of ~0.25deg since 1981-09-01
}
\examples{
load_oisst("1981-09-01", "1981-09-01")
}
