% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rain_dy.R
\name{load_rain_dy}
\alias{load_rain_dy}
\title{Download daily rainfall images}
\usage{
load_rain_dy(lan = "en", listfail = F)
}
\arguments{
\item{lan}{Language of descriptive text within the pictures. Only accepts "e"/ "c" as char.}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download from HKO.
Stores them in the working directory orderly according to year.
}
\examples{
load_rain_dy()
}
