% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rain_hr.R
\name{load_rain_hr}
\alias{load_rain_hr}
\title{Download hourly rainfall images}
\usage{
load_rain_hr(
  ETime = Sys.time(),
  DDays = 7,
  STime = NA,
  lan = "en",
  listfail = F
)
}
\arguments{
\item{ETime}{The newest png to be downloaded, starting from the latest time. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{DDays}{The duration of images to be downloaded in days. Only accepts numerical values.}

\item{STime}{The oldest png to be downloaded, ending at the earliest time. Only accepts POSIXct (tip: create POSIXct via ISOdatetime)}

\item{lan}{Language of the descriptive text}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download from HKO.
Stores them in the working directory orderly according to date.
}
\examples{
load_rain_hr()
}
