% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_satl.R
\name{load_satl}
\alias{load_satl}
\title{Downloads satellite images (v2)}
\usage{
load_satl(
  ETime = Sys.time(),
  DDays = 3.5,
  STime = NA,
  magn = 8,
  type = "dc",
  listfail = F
)
}
\arguments{
\item{ETime}{The time of the lastest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{DDays}{The duration of images to be downloaded in days. Only accepts numerical values.}

\item{STime}{The time of the earliest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{magn}{The satellite image magnification. Only accepts 2/ 4/ 8}

\item{type}{The type of satellite image, which includes "True color" ("tc"), "Infrared" ("ir"), and "Deep convection" ("dc").}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Downloads satellite images (v2)
}
\examples{
load_satl()
}
