% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tide_csv.R
\name{load_tide_csv}
\alias{load_tide_csv}
\title{Download tidal data}
\usage{
load_tide_csv(
  ETime = Sys.time(),
  DDays = 7,
  STime = NA,
  type = "both",
  lan = "en",
  listfail = F
)
}
\arguments{
\item{ETime}{The time of the lastest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{DDays}{The duration of images to be downloaded in days. Only accepts numerical values.}

\item{STime}{The time of the earliest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{type}{Source of data. Only accepts "hko"/ "md"/ "both"}

\item{lan}{Language of the descriptive text}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Downloads and stores them in the working directory orderly according to date.
}
\examples{
load_tide_csv()
}
