% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tide_pre.R
\name{load_tide_pre}
\alias{load_tide_pre}
\title{Present HKO predicted tide information}
\usage{
load_tide_pre(
  Station = "all",
  ETime = Sys.time(),
  STime = Sys.time(),
  Fullname = T
)
}
\arguments{
\item{Station}{Code of the Tidal Station. Only accepts CCH/ CLK/ CMW/ KCT/ KLW/ LOP/ MWC/ QUB/ SPW/ TAO/ TBT/ TMW/ TPK/ WAG.}

\item{ETime}{The latest time to be presented. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{STime}{The earliest time to be presented. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}
}
\value{

}
\description{
Present HKO predicted tide information
}
\examples{
load_tide_pre()
}
