% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wind.R
\name{load_wind}
\alias{load_wind}
\title{Download wind data as image}
\usage{
load_wind(type = "wind", lan = "en", listfail = F)
}
\arguments{
\item{type}{Type of wind. Only accepts "gust"/ "wind" as char.}

\item{lan}{Language of the descriptive text}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download wind data as image
}
\examples{
load_wind()
}
