% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wxph.R
\name{load_wxph}
\alias{load_wxph}
\title{Download weather images from HKO}
\usage{
load_wxph(
  ETime = Sys.time(),
  DDays = 2.5,
  STime = NA,
  type = "all",
  listfail = F
)
}
\arguments{
\item{ETime}{The time of the lastest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{DDays}{The duration of images to be downloaded in days. Only accepts numerical values.}

\item{STime}{The time of the earliest image. Only accepts POSIXct (tip: create POSIXct via ISOdatetime).}

\item{type}{The abbr of the station, which includes lfs, wlp, elc, kfb, tpk, tm2, tlc, sk2, skg, cwb, cwa, ks2, klt, hko, ic2, cp1, vpb, vpa, gsi, swh, slw, dnl, pe2, cch, cce,lam, wl2, wgl. You may include multiple stations in your download, using c(station1, station2, ...).}

\item{listfail}{List the date/time of failed download attempts. Only accepts T/F/1/0.}
}
\value{

}
\description{
Download weather images from HKO
}
\examples{
load_wxph()
}
