% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_create.R
\name{obj_create}
\alias{obj_create}
\title{Creates an hkweather obj from source files}
\usage{
obj_create(rDIR, type)
}
\arguments{
\item{rDIR}{Reading Directory. Which folder should I read to get all the information?}

\item{type}{Type of data. Which type of data it is? "tide" or "wind"}
}
\value{

}
\description{
Creates an hkweather obj from source files
}
\examples{
obj_create(getwd(), type = "tide")
}
