% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_graph.R
\name{obj_graph}
\alias{obj_graph}
\title{Graph}
\usage{
obj_graph(obj, graph = c("point", "violin"))
}
\arguments{
\item{obj}{a hkweather object}

\item{graph}{Type of graph you want to plot. Is it a "point", "violin", or "bar"}
}
\value{

}
\description{
Graph
}
\examples{
obj_graph(obj)
}
