% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_join.R
\name{obj_join}
\alias{obj_join}
\title{Joins 2 hkweather object together}
\usage{
obj_join(obj1, obj2)
}
\arguments{
\item{obj1}{a hkweather obj}

\item{obj2}{a hkweather obj}
}
\value{

}
\description{
Joins 2 hkweather object together
}
\examples{
obj_join(obj1, obj2)
}
