% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_pivot.R
\name{obj_pivot}
\alias{obj_pivot}
\title{Pivot data of an object}
\usage{
obj_pivot(obj)
}
\arguments{
\item{obj}{a hkweather object}
}
\value{

}
\description{
Only pivots data with a single station and a single data type
}
\examples{
obj_pivot(obj)
}
