% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_read.R
\name{obj_read}
\alias{obj_read}
\title{Create an hkweather object from processed file and dataframe}
\usage{
obj_read(rDIR, type, read = "df")
}
\arguments{
\item{rDIR}{Reading Directory. Which folder should I read to get all the information?}

\item{type}{Type of data. Which type of data it is? (Only accepts gtmp/ mslp/ rain/ rhum/ sart/ temp/ tide/ wind)}

\item{read}{Data format of processed data. (Only accepts df/ xlsx/ csv)}
}
\value{

}
\description{
Create an hkweather object from processed file and dataframe
}
\examples{
obj_read(load_tide_pre(), type = "tide", read = "df")
}
