% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_table.R
\name{obj_table}
\alias{obj_table}
\title{Orders everything a table}
\usage{
obj_table(obj, split = c("type", "both"))
}
\arguments{
\item{obj}{a hkweather obj}

\item{split}{how should the data be split? By "type", "station", or "both"?}
}
\value{

}
\description{
Orders everything a table
}
\examples{
table("1/1/1", "tide", write = "xlsx")
}
