% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_view_legend.R
\name{obj_view_legend}
\alias{obj_view_legend}
\title{View the legend of object}
\usage{
obj_view_legend(obj)
}
\arguments{
\item{obj}{a hkweather obj}
}
\value{

}
\description{
View the legend of object
}
\examples{
obj_view_legend(obj)
}
