% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_view_plot.R
\name{obj_view_plot}
\alias{obj_view_plot}
\title{View the plots of an hkweather obj}
\usage{
obj_view_plot(obj, type = "all", index = 1)
}
\arguments{
\item{obj}{an hkweather obj}

\item{type}{type of plot to show, only accepts}

\item{index}{index}
}
\value{

}
\description{
View the plots of an hkweather obj
}
\examples{
obj_view_plot(obj, index = 2)
}
