% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_write.R
\name{obj_write}
\alias{obj_write}
\title{Write data of an object as file}
\usage{
obj_write(obj, wDIR, type = "xlsx")
}
\arguments{
\item{obj}{a hkweather object}

\item{type}{}
}
\value{

}
\description{
Write data of an object as file
}
\examples{
obj_write(obj)
}
