% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_gph.R
\name{tide_gph}
\alias{tide_gph}
\title{Graphs HKOtide and MDtide results}
\usage{
tide_gph(df = NA)
}
\arguments{
\item{df}{dataframe of a "tidy" MD and HKO tide results}
}
\value{

}
\description{
Graphs HKOtide and MDtide results
}
\examples{
tide_gph(x)
}
