% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_pre.R
\docType{data}
\name{tide_pre}
\alias{tide_pre}
\title{Predicted Tide Data of Hong Kong from 2015 to 2024}
\format{
A dataframe with 5 variables
\describe{
\item{Station}{Station}
\item{Hour}{Hour}
\item{Value}{Value}
\item{DateTime}{DateTime}
}

@source {HKO}

@examples
data(tide_pre)
}
\usage{
tide_pre
}
\description{
Contains data of all date and time
}
\keyword{datasets}
