% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_pre_station.R
\docType{data}
\name{tide_pre_station}
\alias{tide_pre_station}
\title{Stations of Predicted Tide Data of Hong Kong from 2015 to 2024}
\format{
A list of stations
\describe{
\item{Station}{Station}
}

@source {HKO}

@examples
data(tide_pre_station)
}
\usage{
tide_pre_station
}
\description{
Contains all stations
}
\keyword{datasets}
