% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_tbl.R
\name{tide_tbl}
\alias{tide_tbl}
\title{Orders HKOtide results (v2)}
\usage{
tide_tbl(DIR, write = T)
}
\arguments{
\item{DIR}{Path of the directory. Accepts list of directory.}

\item{write}{Writes CSV file to Working Directory.}
}
\value{

}
\description{
Tidies up HKOtide csv files. Creates a list of files and names
Can create csv in working directory with default name.
}
\examples{
tide_tbl(getwd())
}
