% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidemd_tbl.R
\name{tidemd_tbl}
\alias{tidemd_tbl}
\title{Orders MDtide results (v2)}
\usage{
tidemd_tbl(DIR, write = T)
}
\arguments{
\item{DIR}{Path of the directory. Accepts list of directory.}

\item{write}{Writes CSV file to Working Directory.}
}
\value{

}
\description{
Tidies up MDtide csv files. Creates a list of files and names
Can create csv in working directory with default name.
}
\examples{
tidemd_tbl(getwd())
}
