% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wthall.R
\name{wthall}
\alias{wthall}
\title{Downloads all the satellite images and lighting images from the Hong Kong Observatory. If there is a Chinese and English version, only downloads the English version.}
\usage{
wthall(listfail = F)
}
\value{

}
\description{
Downloads all the satellite images and lighting images from the Hong Kong Observatory. If there is a Chinese and English version, only downloads the English version.
}
\examples{
wthall()
}
