% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_asnd_plot.R
\name{draw_asnd_plot}
\alias{draw_asnd_plot}
\title{Draw atmospheric sounding plot}
\usage{
draw_asnd_plot(
  data = NA,
  DABT = T,
  MABT = T,
  IHUM = T,
  PRES = c(1000, 100),
  TEMP = c(-100, 50)
)
}
\arguments{
\item{data}{Data from 1 station. Only accepts a dataframe that includes column PRES, TEMP, and DWPT.}

\item{DABT}{Draws the dry adiabatic line. Default as T}

\item{MABT}{Draws the moist adiabatic line. Default as T.}

\item{IHUM}{Draws the isohume line. Default as T.}

\item{PRES}{Pressure range of the atmospheric sounding plot.}

\item{TEMP}{Temperature range of the atmospheric sounding plot.}
}
\value{

}
\description{
Draw atmospheric sounding plot
}
\examples{
draw_asnd_plot()
}
