% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HKOall.R
\name{HKOall}
\alias{HKOall}
\title{Downloads all the satellite images and lighting images from the Hong Kong Observatory. No parameters are utilized}
\usage{
HKOall()
}
\value{

}
\description{
Downloads all the satellite images and lighting images from the Hong Kong Observatory. No parameters are utilized
}
\examples{
HKOall()
}
