% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_graph.R
\name{obj_graph}
\alias{obj_graph}
\title{Graph all the data within an hkweather object}
\usage{
obj_graph(obj, graph = "all")
}
\arguments{
\item{obj}{a hkweather object}

\item{graph}{Type of graph you want to plot. Is it a "point", "violin", or "bar"}
}
\value{

}
\description{
Graph all the data within an hkweather object
}
\examples{
obj_graph(obj)
}
